server/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
PyMuPDF==1.24.9
numpy==2.3.3
google-generativeai==0.8.3
python-docx==1.1.2
pillow==10.4.0
opencv-python==4.10.0.84
gtts==2.5.1
pydub==0.25.1
websockets>=9.0,<12.0
solana==0.34.0
base58==2.0.1
motor==3.3.2
pymongo==4.6.1